FIXUP: addresses most feedback; need to fix GN
