+++ chromium_src/extensions/common/permissions/DEPS
@@ -1,3 +1,3 @@
 include_rules = [
-  "+brave/components/skus/renderer",
+  "+brave/components/skus/common",
 ]
